preprocessor/requirements.txt
pydantic
google-cloud-language
flask
kadia_common==0.0.37
